+++ tests/additive.rs
@@ -1,5 +1,6 @@
 use scl_rs::{
     math::{field::mersenne61::Mersenne61, ring::Ring},
+    net::PartyId,
     ss::additive::AdditiveSS,
 };
 
@@ -10,7 +11,8 @@ fn construct_shares_and_reconstruct() {
 
     for _ in 0..REPETITIONS {
         let secret = Mersenne61::random(&mut rand::rng());
-        let shares = AdditiveSS::shares_from_secret(secret, N_PARTIES, &mut rand::rng());
+        let parties: Vec<PartyId> = (0..N_PARTIES).map(PartyId::from).collect();
+        let shares = AdditiveSS::shares_from_secret(secret, &parties, &mut rand::rng());
         let rec_secret = AdditiveSS::secret_from_shares(&shares);
         assert_eq!(secret, rec_secret);
     }